fix: index files added
