+++ src/app/core/constants/index.ts
@@ -1,0 +1,2 @@
+export * from './menu';
+export * from './routes';
