+++ src/app/core/index.ts
@@ -1,0 +1,2 @@
+export * from '../mocks';
+export * from './constants';
