+++ src/app/mocks/index.ts
@@ -1,15 +1,1 @@
-export const mockNotifications = [
-    {
-        name: 'User1',
-        desc: [
-            { text: 'Bla-bla-bla Bla-bla-bla Bla-bla-bla Bla-bla-bla', date: new Date() },
-            { text: 'La-la-la La-la-la La-la-la', date: new Date() },
-        ],
-    },
-];
-
-export const mockUser = {
-    name: 'mockUser',
-    surname: 'Surr',
-    role: 'admin',
-};
+export * from './mocks';
