+++ src/app/mocks/mocks.ts
@@ -1,0 +1,15 @@
+export const mockNotifications = [
+    {
+        name: 'User1',
+        desc: [
+            { text: 'Bla-bla-bla Bla-bla-bla Bla-bla-bla Bla-bla-bla', date: new Date() },
+            { text: 'La-la-la La-la-la La-la-la', date: new Date() },
+        ],
+    },
+];
+
+export const mockUser = {
+    name: 'mockUser',
+    surname: 'Surr',
+    role: 'admin',
+};
